no need for this
If you run the following command, you can install a symbolic link
package of your development directory so that there is no need to
explicitly point to the development directory within the code. This is
prefereable so that the example code should work without the need to
have the development directory available.